El muro - borra comentarios

+++ app/views.py
@@ -7,7 +7,6 @@ from datetime import datetime, time, timedelta
 from django.utils import timezone
 import re
 
-
 @login_required
 def index(request):
     context = {
@@ -37,15 +36,6 @@ def mensaje_crear(request):
         messages.success(request, "Mensaje agregado correctamente")
         return redirect(f"/")
 
-
-"""@login_required
-def mensaje_borrar(request, val):
-    print(request.GET, 'Entró a borrar mensaje')
-    borr = Mensaje.objects.get(id = val)
-    print("Aqui se va a borrar el mensaje ID=", val)
-    borr.delete()
-    return redirect("/")"""
-
 def calculate_minutos(fecha):
     today = timezone.now()
     print(today, "today")
@@ -55,10 +45,8 @@ def calculate_minutos(fecha):
     print("dias", (today.day - fecha.day), (today.day - fecha.day)*24*60)
     print("horas", (today.hour-fecha.hour), (today.hour-fecha.hour)*60)
     print("minutos", (today.minute-fecha.minute))
-
     resultado = (today.year - fecha.year)*365*24*60 + (today.month - fecha.month)*30*24*60 + (today.day - fecha.day)*24*60 + (today.hour-fecha.hour)*60 + (today.minute-fecha.minute)
     print(resultado)
-    #print(today.datetime()-fecha.datetime())
     return resultado
 
 
@@ -98,15 +86,18 @@ def comentario_crear(request):
             mensaje = Mensaje.objects.get(id=id_mensaje),
         )
         messages.success(request, "Comentario agregado correctamente")
-
         return redirect(f"/")
-
 
 @login_required
 def comentario_borrar(request, val):
     print(request.GET)
     borr = Comentario.objects.get(id = val)
     print("Aqui se va a borrar el comentario ID=", val)
-    borr.delete()
-    
+    calc = calculate_minutos(borr.created_at)
+    if calc > 30:
+        messages.warning(request, "No se borró el comentario. Tiempo expirado máximo 30 min")
+    else: 
+        print("Autorizado para borrar el comentario")
+        messages.success(request, "Comentario borrado exitosamente")
+        borr.delete()
     return redirect("/")
